Modify import dotfiles

+++ config.sh
@@ -20,11 +20,11 @@ EOF
 # install grub and generate config file
 if [ -d /sys/firmware/efi/efivars ]; then
     grub-install --target=x86_64-efi --efi-directory=/efi --boot-directory=/efi --bootloader-id=GRUB > /dev/null 2>&1
-    sed -i "s/GRUB_TIMEOUT=5/GRUB_TIMEOUT=1/g" /etc/default/grub
+    sed -i "s/GRUB_TIMEOUT=5/GRUB_TIMEOUT=0/g" /etc/default/grub
     grub-mkconfig --output=/efi/grub/grub.cfg > /dev/null 2>&1
 else
     grub-install --target=i386-pc /dev/sda > /dev/null 2>&1
-    sed -i "s/GRUB_TIMEOUT=5/GRUB_TIMEOUT=1/g" /etc/default/grub
+    sed -i "s/GRUB_TIMEOUT=5/GRUB_TIMEOUT=0/g" /etc/default/grub
     grub-mkconfig --output=/boot/grub/grub.cfg > /dev/null 2>&1
 fi
 
@@ -35,9 +35,10 @@ echo "root:$3" | chpasswd
 sed -i "s/^# %wheel ALL=(ALL) NOPASSWD: ALL/%wheel ALL= NOPASSWD: ALL, NOPASSWD: \/usr\/bin\/halt, \/usr\/bin\/poweroff, \/usr\/bin\/reboot, \/usr\/bin\/pacman -Syu, \/usr\/bin\/pacman -Syyu, \/usr\/bin\/pacman -Sy, \/usr\/bin\/pacman -Syy/g; /without a password/a Defaults \!tty_tickets" /etc/sudoers
 
 # import dotfiles
-git clone -q https://github.com/ifananvity/dotfiles.git &&
-mv -f dotfiles/.* $HOME
-rm -rf dotfiles
+git clone -q https://github.com/ifananvity/dotfiles.git > /dev/null 2>&1
+for dot in $(ls dotfiles/ -a); do
+	[ ! $dot = ".." ] && [ ! $dot = "." ] && mv -f $dot /home/$4/
+done && rm -rf dotfiles/
 
 # colorize output
 sed -i "s/^#Color/Color/g; /#VerbosePkgLists/a ILoveCandy" /etc/pacman.conf
